wasi: stops defaulting to os.Stdin,out,err
Defaulting to os.Stdin os.Stdout and os.Stderr doesn't make sense for
the same reasons as why we don't propagate ENV or ARGV: it violates
sand-boxing. Moreover, these are worse as they prevent concurrency and
can also lead to console overload if accidentally not overridden.

This PR defaults writers to discards and stdin to err. Erring on reads
from an unconfigured os.Stdin is a better choice than trying to use
`os.DevNull` as the latter requires management of a file pointing to it
(ex having to open and close that which is complicated when many modules
import one instance of WASIConfig).

Signed-off-by: Adrian Cole <adrian@tetrate.io>